refactor: move code to new components
